requirements.txt
Django==3.2.11
django-ckeditor==6.2.0
six==1.16.0
openpyxl==3.0.9
Unidecode==1.3.2
Pillow==8.4.0
django-bootstrap4==21.1
django-bootstrap-datepicker-plus==4.0.0
django-dbshell-plus==1.0.1
django-uuslug==1.2.0
requests==2.28.1
djangorestframework==3.14.0